Split search result helper components
