+++ src/modules/search/components/SearchResultsEmptyState.tsx
@@ -1,0 +1,9 @@
+import { CommandEmpty } from "@components/ui/command";
+
+type SearchResultsEmptyStateProps = {
+  message: string;
+};
+
+export const SearchResultsEmptyState = ({
+  message,
+}: SearchResultsEmptyStateProps) => <CommandEmpty>{message}</CommandEmpty>;
