+++ src/modules/search/components/SearchResultsErrorState.tsx
@@ -1,0 +1,7 @@
+import { CommandEmpty } from "@components/ui/command";
+
+export const SearchResultsErrorState = () => (
+  <CommandEmpty>
+    Search is unavailable. Please try again in a moment.
+  </CommandEmpty>
+);
