+++ src/modules/search/components/SearchResultsLoadingState.tsx
@@ -1,0 +1,9 @@
+import { CommandLoading } from "cmdk";
+
+import { SearchResultSkeletonRows } from "./SearchResultSkeletonRows";
+
+export const SearchResultsLoadingState = () => (
+  <CommandLoading>
+    <SearchResultSkeletonRows />
+  </CommandLoading>
+);
